Tested and tweaked all tests to run smoothly on all platforms.
Also removed the entire task relating to updating the load method
because teaching JSON.NET feels out of place and the alternative made
it impossible to make the class abstract. Instead I created a very much
not in line with best practices conversion method capable of converting
json into not yet existing objects using reflection. (The offending
conversion method is specifically noted as not a great idea, but a
necessity to avoid complicating the project for students)